requirements.txt
google-generativeai
gitpython
rich
python-dotenv
